app/requirements.txt
# flake8<4.0
# flask>=1.0,<1.1
# nose<2.0
# pytesseract>=0.2.1,<0.3

# pillow>=6.2.3 # not directly required, pinned by Snyk to avoid a vulnerability
# jinja2
flake8
flask
numpy
nose
pytesseract
pillow
jinja2
flask_restx
pdf2image
werkzeug==2.1.2
